requirements.txt
pydantic==1.8.2
requests==2.25.1
python-dotenv==0.17.1
pandas==1.2.4
